The previous commit to master was making up for mistakenly deleting the local repository and having to merge a new local repository with the old remote.  Although it makes the repository somewhat dirty, the only way I found to do that was copy all old commits into a branch (v1) and then commit the new repository to master.  So all commits prior to this point can be found in the v1 branch.
As for the code, the source and doublet influence formulations have been changed from the Katz and Plotkin formulation to the VSAERO formulation.  This avoids a coordinate transformation, making it faster and easier to follow.  The code has been validated against the analytical solution for flow around an ellipsoid, although the velocity survey does not provide very accurate velocities due to the singularity as you approach the vortex making up the edges of doublet panels.  A finite core model was applied to make the induced velocity approach zero smoothly, making the velocity survey more accurate but for precise velocities, the derivative of the potential will be calculated using a least squares fit of the surrounding panels.  This code is next in the development process.